[US_ND] Use facility to prioritize cases in incarceration_staff_sessions (Recidiviz/recidiviz-data#39847)
## Description of the change

We found a case of a person assigned to two people: a CM who was in his
facility (and therefore it made sense) and a CM who was not in this
person's facility. We ended up assigning this person to the CM that did
not make sense because the latter had a later assignment date, which we
used to determine priority. Now we included one additional column to
determine priority `jii_facility_matches_staff` which checks if the
staff has been in the same facility as the person in a given assignment
session. This is now fixing the issue.

I uploaded this into a sandbox and it all looks well. There's now 18
people in the resident record who have changed their assignment and 3
who had a NULL assignment who now have an actual `officer_id`.

```
python -m recidiviz.tools.load_views_to_sandbox \
        --project_id recidiviz-staging\
        --sandbox_dataset_prefix hsalas \
        --state_code_filter US_ND manual \
        --view_ids_to_load sessions.us_nd_incarceration_staff_assignment_sessions_preprocessed,sessions.incarceration_staff_assignment_sessions_preprocessed,workflows_views.resident_record\
        --update_descendants False
```

```
WITH sandbox AS (
  SELECT state_code, person_external_id, officer_id AS officer_id_sandbox,
  FROM `recidiviz-staging.hsalas_workflows_views.resident_record_materialized`
),

main AS (
  SELECT state_code, person_external_id, officer_id AS officer_id_main,
  FROM `recidiviz-staging.workflows_views.resident_record_materialized`
  WHERE state_code = 'US_ND'
)

SELECT *
FROM main
LEFT JOIN sandbox
  USING(state_code, person_external_id)
WHERE IFNULL(officer_id_sandbox, '') != IFNULL(officer_id_main, '')
```

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 80d748dd7285b92fbe71a6fe4974395119d941f7