[US_IA] Ingest sentences and charges (Recidiviz/recidiviz-data#37077)
## Description of the change

This PR ingests sentences, sentence groups, sentence lengths, sentence
group lengths, and sentence status snapshots for IA.

### Details:

In IA, the data for charges and sentences come in at three levels:
- At the charge level, which contains information about the
offense/charge, and corresponds with a single "case"
- At the sentence level, where there might be multiple sentences
associated with a single charge. This level includes an imposed date and
an end date.
- At the penalty level, where there might be multiple penalties
associated with a single sentence. This level the type of penalty
imposed (i.e. fine vs jail vs prison vs probation etc.)

Therefore, for ingest, I've considered each charge to be a
StateSentenceGroup and each penalty to be a StateSentence. All penalties
associated with the same sentence will have the same imposed date and
end date when we ingest them as StateSentences, since those dates only
live on the "sentence" level in the raw data.

In addition, Iowa does not impose probation sentences directly. Instead,
they impose a prison sentence which they then suspend with probation.
Therefore, in ingest, we set sentence type to PROBATION whenever we see
any sort of sentence suspended with probation.

There are some outstanding data questions / a few other areas to flesh
out in ingest, but since they are unneeded at this time, I have created
a ticket to address them later: Recidiviz/recidiviz-data#38402

### Testing
- [x] Successful ingest sandbox

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 034fd718266025c309dae66d534515857382340e