[US_ND] Protect against nulls in sentencing staff record (Recidiviz/recidiviz-data#39926)
## Description of the change

The `docstars_psi` table came in with a malformed row that listed the
PSI's name as "FirstName LastName, FirstName LastName" rather than the
standard "LastName, FirstName" that we have built our logic around. This
PR adds a raw data migration to fix the row so that the name is in the
expected format.

The fallout of this was that there was a row in the sentencing staff
record with an assigned client, but only NULLs in fields that contain
information about the actual staff. This PR also changes the join from
cases to staff in the staff record to an `INNER JOIN` to avoid having
that happen again. A possible pitfall of this is that we will not know
when we are excluding cases because of a mismatch on staff name (for
instance, we couldn't have identified this problem if this `INNER JOIN`
had already been in place)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 392acbbb07a9d03e4b1a9865b94ff21ceead1e63