[Utils] Allow for custom start date field name in create_sub_sessions_with_attributes (Recidiviz/recidiviz-data#39948)
## Description of the change

I want to be able to use the `create_sub_sessions_with_attributes`
helper in an ingest view without having to change the names of the raw
data columns it sub-sessionizes to `start_date` and `end_date`. We
already have a way to set a custom end date column - this just updates
the helper to also allow for a custom start date.

To test this, I rebased on the latest alpha deploy and confirmed there
were no view updates from this change.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Unblocks some work for Recidiviz/recidiviz-data#38347

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2e60d1baadbf031bff8daceb9369e27393c07a3d