[US_AZ] Remove extra rows from agreement form view (Recidiviz/recidiviz-data#39993)
## Description of the change

We were pulling in extra rows with no information about agreement form
signature status only in the TPR-related CTE of this view. This was
making it so that once a form was signed, there could be subsequent
updates that suggested it had not yet been signed. This takes away the
COALESCE that was creating those inaccurate rows and makes the logic in
the TPR and DTP CTEs match.

I loaded this view to a sandbox and it fixed the one example error we
have.

Current:
```
SELECT * FROM recidiviz-staging.analyst_data.us_az_agreement_form_signatures d
WHERE person_id = 450128989122380856
```

Corrected:
```
SELECT * FROM recidiviz-staging.elise_analyst_data.us_az_agreement_form_signatures d
WHERE person_id = 450128989122380856

```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 34a95cb582df7164758b620d722af5dcdca23844